+++ changelog.md
@@ -47,6 +47,7 @@
 - Change Item Permission to check in all inventory instead of just the player's hand
 - Create new Button for zMenu+ ``INPUT``, allows you to choose a number or a text in the chat and perform actions
 - You can choose the pattern plugin for the button with ``pluginName``
+- Fix MenuItemStackLoader for save model id
 
 # 1.0.3.0